# docs/swagger.yaml
basePath: /
definitions:
  http.CreateDepartmentRequest:
    properties:
      name:
        type: string
      parent_id:
        type: integer
    type: object
  http.CreateDepartmentResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
    type: object
  http.CreateEmployeeRequest:
    properties:
      full_name:
        type: string
      hired_at:
        type: string
      position:
        type: string
    type: object
  http.CreateEmployeeResponse:
    properties:
      created_at:
        type: string
      department_id:
        type: integer
      full_name:
        type: string
      hired_at:
        type: string
      id:
        type: integer
      position:
        type: string
    type: object
  http.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  http.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/http.Error'
    type: object
  http.UpdateDepartmentRequest:
    properties:
      name:
        type: string
      parent_id:
        type: integer
    type: object
  http.UpdateDepartmentResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
    type: object
  query.Department:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
    type: object
  query.Employee:
    properties:
      created_at:
        type: string
      department_id:
        type: integer
      full_name:
        type: string
      hired_at:
        type: string
      id:
        type: integer
      position:
        type: string
    type: object
  query.GetDepartment:
    properties:
      children:
        items:
          $ref: '#/definitions/query.Department'
        type: array
      department:
        $ref: '#/definitions/query.Department'
      employees:
        items:
          $ref: '#/definitions/query.Employee'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  title: Departments API
  version: "1.0"
paths:
  /departments/:
    post:
      consumes:
      - application/json
      description: Creates a new department, optionally nested under a parent
      parameters:
      - description: Department payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.CreateDepartmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.CreateDepartmentResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Department parent not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Department name conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create a department
      tags:
      - departments
  /departments/{id}:
    delete:
      description: |-
        Deletes a department using one of two modes:
        - `cascade` — deletes the department along with all its employees and sub-departments recursively.
        - `reassign` — deletes the department and moves its employees to the department specified by `reassign_to_department_id`.
        Note: `reassign_to_department_id` is required when mode is `reassign`.
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Deletion mode
        enum:
        - cascade
        - reassign
        in: query
        name: mode
        required: true
        type: string
      - description: Target department ID for employee reassignment (required when
          mode=reassign)
        in: query
        name: reassign_to_department_id
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid mode or missing reassign_to_department_id
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Department not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Delete a department
      tags:
      - departments
    get:
      description: Returns a department with its employees and nested sub-departments
        up to the specified depth
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 'Depth of nested departments (default: 1, max: 5)'
        in: query
        maximum: 5
        minimum: 1
        name: depth
        type: integer
      - default: true
        description: 'Include employees in response (default: true)'
        in: query
        name: include_employees
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/query.GetDepartment'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Department not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get department details with employees and subtree
      tags:
      - departments
    patch:
      consumes:
      - application/json
      description: Updates department name and/or moves it under a new parent. All
        body fields are optional.
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.UpdateDepartmentResponse'
        "400":
          description: Invalid request body or department ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Department not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Circular parent reference detected
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Update a department
      tags:
      - departments
  /departments/{id}/employees/:
    post:
      consumes:
      - application/json
      description: Creates a new employee and assigns them to the specified department
      parameters:
      - description: Department ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employee payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.CreateEmployeeResponse'
        "400":
          description: Invalid request body or department ID
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Department not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create an employee in a department
      tags:
      - employees
swagger: "2.0"